增加token，权限校验

+++ server/controllers/person.js
@@ -1,5 +1,5 @@
 const Mperson = require("../model/person");
-const { SUCESS_RES, ERROR_RES } = require("../utils/resData");
+const response = require("../utils/resData");
 
 // 书摘的接口----------------------------------------------------------------
 // 构造函数写法
@@ -7,27 +7,29 @@ function Person() {}
 // 资源共享，节约内存
 Person.prototype.findExcerpt = async function (ctx, next) {
   try {
-    ctx.body = SUCESS_RES.getCode(await Mperson.findAll());
+    ctx.body = response.SUCCESS("common", await Mperson.findAll());
   } catch (error) {
     console.log(error);
-    ctx.body = ERROR_RES.getCode(null);
+    ctx.body = response.SERVER_ERROR();
   }
 };
 Person.prototype.addExcerpt = async function (ctx, next) {
   try {
     const { content, author, flag, date } = ctx.request.query;
-    ctx.body = SUCESS_RES.getCode(
+    ctx.body = response.SUCCESS(
+      "common",
       await Mperson.create({ content, author, flag, date })
     );
   } catch (error) {
     console.log(error);
-    ctx.body = ERROR_RES.getCode(null);
+    ctx.body = response.SERVER_ERROR();
   }
 };
 Person.prototype.updateExcerpt = async function (ctx, next) {
   try {
     const { id, content } = ctx.request.query;
-    ctx.body = SUCESS_RES.getCode(
+    ctx.body = response.SUCCESS(
+      "common",
       await Mperson.update(
         { content: content },
         {
@@ -39,12 +41,17 @@ Person.prototype.updateExcerpt = async function (ctx, next) {
     );
   } catch (error) {
     console.log(error);
-    ctx.body = ERROR_RES.getCode(null);
+    ctx.body = response.SERVER_ERROR();
   }
 };
 Person.prototype.delExcerpt = async function (ctx, next) {
   try {
-    ctx.body = SUCESS_RES.getCode(
+    if (!ctx.state.user.is_admin) {
+      ctx.body = response.ERROR("powerLacking");
+      return;
+    }
+    ctx.body = response.SUCCESS(
+      "common",
       await Mperson.destroy({
         where: {
           id: ctx.request.query.id,
@@ -53,7 +60,7 @@ Person.prototype.delExcerpt = async function (ctx, next) {
     );
   } catch (error) {
     console.log(error);
-    ctx.body = ERROR_RES.getCode(null);
+    ctx.body = response.SERVER_ERROR();
   }
 };
